seewhy-live/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-socketio==5.11.4
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
redis[hiredis]==5.2.1
celery[redis]==5.4.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
pydantic==2.10.4
pydantic-settings==2.7.0
httpx==0.28.1
anthropic==0.42.0
stripe==11.4.1
pillow==11.0.0
boto3==1.35.91
aiofiles==24.1.0
slowapi==0.1.9
email-validator==2.2.0
